backend/requirements.txt
numpy==1.26.4
matplotlib==3.9.2
nltk==3.9.1
torch>=2.5.1
transformers>=4.16
pandas>=1.3.0
scikit-learn>=0.24.0
Django==5.1.3
djangorestframework>=3.15.2
datasets>=3.1.0
lime==0.2.0.1
accelerate>=0.26.0
